+++ lib/braintree/payment_method_gateway.rb
@@ -47,8 +47,6 @@ module Braintree
         SuccessfulResult.new(:payment_method => CoinbaseAccount._new(@gateway, response[:coinbase_account]))
       elsif response.has_key?(:us_bank_account)
         UsBankAccount._new(@gateway, response[:us_bank_account])
-      elsif response.has_key?(:europe_bank_account)
-        EuropeBankAccount._new(@gateway, response[:europe_bank_account])
       elsif response.has_key?(:apple_pay_card)
         ApplePayCard._new(@gateway, response[:apple_pay_card])
       elsif response.has_key?(:android_pay_card)